Calculates how long the bulb has been turned on
There is a light bulb, which by default is off, and a button, by pressing which the light bulb switches its state. This means that if the light bulb is off and the button is pressed, the light turns on, and if you press it again, it turns off.

Input: A list of datetime objects

Output: A number of seconds as an integer.