Determine how long the light bulb has been turned on
There is a light bulb, which by default is off, and a button, by pressing which the light bulb switches its state. This means that if the light bulb is off and the button is pressed, the light turns on, and if you press it again, it turns off. I want not only to know how long there was light in the room, but how long the room was lit, starting from a certain moment.